+++ public/stylesheets/logs.css
@@ -23,7 +23,6 @@ table {
     minmax(150px, 1.67fr)
     minmax(150px, 0.5fr)
     minmax(150px, 0.5fr)
-    minmax(150px, 0.5fr)
     minmax(150px, 3fr)
     minmax(150px, 0.5fr)
     minmax(150px, 5.33fr);
